All client methods in methods.go verify the response http status code
While most endpoints unmarshal json, which will error on unexpected
responses, verifying the status code first is less expensive, and
provides more accurate and descriptive failure messages in the logs.

I have verified each endpoints response statuses in the official
qbittorrent documentation, with the exception of `torrents/export`,
which is not documented. For this endpoint I have implemented
specific handling for 404 not found, but left the 409 conflict status
to fall through to the default unknown status branch, based on the
qbittorrent implementation:
https://github.com/qbittorrent/qBittorrent/blob/a77b17e6dafadca1eb0f80d4c0167c791107f787/src/webui/api/torrentscontroller.cpp#L2045-L2059
The qbittorrent 409 conflict status code seems questionable, as the
failure appers not to be related to the request, so it would
make more sense as a 5xx rather than 4xx status class, and I wouldn't
be surprised if this changed in the future.

This is a follow-up to #97 which handles the rest of the endpoints.